% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city_list.R
\docType{data}
\name{city_list}
\alias{city_list}
\title{city_list}
\format{An object of class \code{data.frame} with 209579 rows and 4 columns.}
\usage{
city_list
}
\description{
This dataset contains city names, related country and the coordinates of the
city.
}
\keyword{dataset}
