% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cityweather.R
\docType{class}
\name{cityweather-class}
\alias{cityweather-class}
\alias{cityweather}
\title{cityweather}
\description{
This object can connect to the openweathermap API and get the 7 days
forecasting of the desired city. For more information about this API
see the link bellow :
\url{https://openweathermap.org/}
}
\section{Fields}{

\describe{
\item{\code{cityname}}{A character shows the city name.}

\item{\code{content}}{A list contains the content of API response.}

\item{\code{status_code}}{An integer that shows the status code of API response.}

\item{\code{has_error}}{A boolean that shows TRUE if there is an error.}
}}

\section{Methods}{

\describe{
\item{\code{initialize(cityname, key = "3c656bd3014279a8f41b90522c014977")}}{Getting the data from API}

\item{\code{print()}}{Print out the structure of response content}
}}

