% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weatherforecast.R
\docType{package}
\name{weatherforecast}
\alias{weatherforecast}
\alias{weatherforecast-package}
\title{weatherforecast}
\description{
This package is a R client for OpenWeatherMap API and it can be used to get
the 5 days forecast for the cities around the world.
}
\author{
Nahid Farazmand
}
